B17 (#20)
<!-- Thanks for sending a pull request! Please check out our
contribution guidelines: -->
<!-- https://github.com/hyperledger/besu/blob/main/CONTRIBUTING.md -->

## PR description
Besu pipeline is a sequence of stages and each stage, gets input data
from the previous stage, can be executed in parallel to other stages,
and put the executions results (output) into the input of the next
stage.

We have for some tracing endpoints Out Of Memory errors because Besu
keeps in memory all the traces. The pipeline will help in this case to
break down the whole process into steps and keep in memory the objects
needed for each step. This is one way to stream the work.

For example, during block tracing, instead of keeping in memory all the
traces, with this PR we'll keep in memory only the traces of few
transactions (the number of transactions is defined in the pipeline and
may change in the future), but we still need to keep in memory, for now,
the object representation of the json file which consumes much less
memory. Other optimisations in the future might improve this part.

Below the current implementation of the pipeline, this implementation
may change in the future to improve performance or better manage java
memory footprint.

<img width="1100" alt="image"
src="https://user-images.githubusercontent.com/5099602/220935843-f0adb851-176f-4c8a-afd9-c59947faf21f.png">


## Fixed Issue(s)
<!-- Please link to fixed issue(s) here using format: fixes #<issue
number> -->
<!-- Example: "fixes #2" -->

## Documentation

- [x] I thought about documentation and added the `doc-change-required`
label to this PR if
[updates are
required](https://wiki.hyperledger.org/display/BESU/Documentation).

## Acceptance Tests (Non Mainnet)

- [x] I have considered running `./gradlew acceptanceTestNonMainnet`
locally if my PR affects non-mainnet modules.

## Changelog

- [ ] I thought about the changelog and included a [changelog update if
required](https://wiki.hyperledger.org/display/BESU/Changelog).

---------

Signed-off-by: Sally MacFarlane <macfarla.github@gmail.com>